refactor: 순환 의존성 문제 해결
