+++ src/main/java/com/hyunn/commerceplatform/security/handler/CustomAuthenticationFailureHandler.java
@@ -2,7 +2,7 @@ package com.hyunn.commerceplatform.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hyunn.commerceplatform.entity.Users;
-import com.hyunn.commerceplatform.service.UserService;
+import com.hyunn.commerceplatform.service.UserLockService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -18,9 +18,9 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationFa
 @RequiredArgsConstructor
 public class CustomAuthenticationFailureHandler extends SimpleUrlAuthenticationFailureHandler {
 
-  private final UserService userService;
+  private final UserLockService userLockService;
 
-  @Value("${app.login.max-attempts}")
+  @Value("${app.login.max-fail-attempts}")
   private int maxAttempts;
 
   @Value("${app.login.lock-duration-minutes}")
@@ -31,28 +31,30 @@ public class CustomAuthenticationFailureHandler extends SimpleUrlAuthenticationF
       AuthenticationException exception) throws IOException {
     String username = request.getParameter("username");
     String errorMessage;
-
+    Users user = userLockService.getUserOrThrow(username);
     if (exception instanceof LockedException) {
+
       errorMessage = String.format("Account is locked. Please try again after %d minutes.",
           lockDurationMinutes);
     } else {
-      Users user = userService.getUserOrThrow(username);
+
       if (user.isAccountNonLocked()) {
         if (user.getFailedAttempt() < maxAttempts - 1) {
-          userService.incrementFailedAttempts(user);
+          userLockService.incrementFailedAttempts(user);
           int attemptsLeft = maxAttempts - user.getFailedAttempt();
           errorMessage = String.format(
               "Invalid credentials. %d attempt%s left before account lockout.",
               attemptsLeft, attemptsLeft == 1 ? "" : "s");
         } else {
-          userService.lockUser(user);
+          userLockService.lockUser(user);
           errorMessage = String.format(
               "Account locked due to %d failed attempts. It will be unlocked after %d minutes.",
               maxAttempts, lockDurationMinutes);
         }
       } else {
-        if (userService.unlockWhenTimeExpired(user)) {
-          errorMessage = "Account has been unlocked. Please try to login again.";
+
+        if (userLockService.unlockWhenTimeExpired(user)) {
+          errorMessage = "Account has been unlocked. Please tr y to login again.";
         } else {
           errorMessage = String.format("Account is still locked. Please try again after %d minutes",
               lockDurationMinutes);
