+++ src/main/resources/application.properties
@@ -47,7 +47,9 @@ springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/swagger-ui.html
 # GeoLite2 Configuration
 geoip.database.path=/path/to/GeoLite2-City.mmdb
-login.log.retention.days=30
-# Login lock Configuration
+# Login Log Configuration
+app.login.log.retention.days=30
+app.login.log.cleanup.cron=0 0 1 * * ?
+# Login Lock Configuration
 app.login.max-fail-attempts=5
 app.login.lock-duration-minutes=5
